Implement differential testing
This patch implements a referenceRunner that executes a script again
if the fuzzer has been started either with `--differentialFuzzing`.

It is a reimplementation of the work done by Lukas Bernhard et al. in
their repository https://github.com/RUB-SysSec/JIT-Picker, but differs
from their work in some aspects. Notable differences:

- Differential Fuzzing is enabled with --differentialFuzzing, there is
  no control over percentage rates.
- This PoC is specifically aimed at the new Maglev JIT compiler.
- A new profile V8MaglevDifferentialFuzzingProfile has been added, which
  borrows a lot from the V8Profile, but also implements custom code
  generators with custom weights. That means, as of now, not every
  generated program actually has calls to `fuzzilli_hash()`. This may
  change in the future.